Added an option for specifying the mipmap level of the image
I added a new command-line option to specify the mipmap level of the
image. The value of the option works the same way as the second parameter
to glTexImage2D, so 0 is the main image and 1 is the first mipmap image,
and so on. By default we use the main image to match the behavior of the
previous version of the tool.