Fixed bug where setting a mipmap level resulted in incorrect dimensions
There was a bug where specifying a mipmap level greater than zero
caused the resulting image dimensions to be incorrect. I fixed it so
that the dimensions (of the extracted image as well as those written
as filename suffices) are of what we actually load. For example, if
you have main image of 128 by 128 pixels and specify mipmap level 1,
the dimensions will become 64 by 64.